backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
python-dotenv==1.0.0

# YAML
python-multipart
llama-index
llama-cloud